backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.4,<3.0
celery>=5.4,<6.0
redis>=5.0,<6.0
python-multipart>=0.0.9,<1.0
pillow>=10.4,<12.0

